Minor refactoring in counterfactual package (#1634)
