+++ libs/counterfactuals/src/util/getCounterfactualChartOptions.ts
